first try

+++ app/src/main/java/com/edu/ulab/app/service/impl/BookServiceImpl.java
@@ -2,35 +2,37 @@ package com.edu.ulab.app.service.impl;
 
 import com.edu.ulab.app.dto.BookDto;
 import com.edu.ulab.app.entity.Book;
+import com.edu.ulab.app.exception.NotFoundException;
 import com.edu.ulab.app.generator.IdGenerator;
-import com.edu.ulab.app.generator.impl.BookIdGeneratorImpl;
+import com.edu.ulab.app.mapper.BookMapper;
 import com.edu.ulab.app.service.BookService;
 import com.edu.ulab.app.storage.BookStorage;
-import com.edu.ulab.app.storage.impl.BookStorageImpl;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
-import java.util.Optional;
+import java.util.List;
 
 @Slf4j
 @Service
 public class BookServiceImpl implements BookService {
 
-    private final IdGenerator idGenerator = new BookIdGeneratorImpl();
+    @Autowired
+    @Qualifier("BookId")
+    private IdGenerator idGenerator;
 
-    private final BookStorage bookStorage = new BookStorageImpl();
+    @Autowired
+    private BookStorage bookStorage;
+
+    @Autowired
+    private BookMapper bookMapper;
 
     @Override
     public BookDto createBook(BookDto bookDto) {
         bookDto.setId(idGenerator.next());
 
-        Book newBook = Book.builder()
-                .id(bookDto.getId())
-                .userId(bookDto.getUserId())
-                .author(bookDto.getAuthor())
-                .title(bookDto.getTitle())
-                .pageCount(bookDto.getPageCount())
-                .build();
+        Book newBook = bookMapper.BookDtoToBook(bookDto);
 
         bookStorage.save(newBook);
 
@@ -38,45 +40,17 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public BookDto updateBook(BookDto bookDto) {
-        Optional<Book> optionalBook = bookStorage.findById(bookDto.getId());
-
-        if(optionalBook.isPresent()) {
-            Book existBook = optionalBook.get();
-
-            existBook.setUserId(bookDto.getUserId());
-            existBook.setTitle(bookDto.getTitle());
-            existBook.setAuthor(bookDto.getAuthor());
-            existBook.setPageCount(bookDto.getPageCount());
-
-            bookStorage.update(existBook);
-
-        }
-
-        return bookDto;
-    }
-
-    @Override
-    public BookDto getBookById(Long id) {
-        Optional<Book> optionalBook = bookStorage.findById(id);
-
-        BookDto bookDto = new BookDto();
-
-        if(optionalBook.isPresent()){
-            Book existBook = optionalBook.get();
-
-            bookDto.setId(id);
-            bookDto.setUserId(existBook.getUserId());
-            bookDto.setTitle(existBook.getTitle());
-            bookDto.setAuthor(existBook.getAuthor());
-            bookDto.setPageCount(existBook.getPageCount());
-        }
-
-        return bookDto;
-    }
-
-    @Override
     public void deleteBookById(Long id) {
+
+        if (bookStorage.findById(id).isEmpty()) {
+            throw new NotFoundException("This book is absent");
+        }
+
         bookStorage.delete(id);
+    }
+
+    @Override
+    public List<Long> getByUserId(Long userId) {
+        return bookStorage.findByUserId(userId);
     }
 }
